refactor: Add modules and adjusted classes
