+++ lib/games.rb
@@ -1,3 +1,4 @@
+
 class Games
   attr_reader :data
