+++ lib/module.rb
@@ -1,0 +1,7 @@
+require_relative './stat_tracker'
+
+module PracticeModule
+
+
+
+end
